SequentialProcessing.m Version 2.0
Followed the steps for Sequential Processing in Week 6 Lab Task: Automated Testing. I added three variable options with different data sizes which loops through each option. At the end, the total time taken to process through each data size is stored into an array. This was done in order to be able to automate testing rather than change the data size each time. The variables can be changed and it will still work the same - makes it a lot easier to complete the testing.