Updates to ParallelProcessing.m Version 3.0
This new and final version includes further functionalities of plotting a graph for each data set automatically, this gets outputted after each data set. The line of best fit and the recommended number of processors is shown after all data sets have been processed. This is so that there would be no manual plotting required by the user.